RegistrationForPreInstalled

+++ aws-greengrass-testing-resources/aws-greengrass-testing-resources-iot/src/main/java/com/aws/greengrass/testing/resources/iot/IotThingModel.java
@@ -32,16 +32,18 @@ interface IotThingModel extends AWSResource<IotClient> {
 
     @Override
     default void remove(IotClient client) {
-        client.listThingPrincipals(ListThingPrincipalsRequest.builder().thingName(thingName()).build()).principals()
-                .forEach(p -> {
-                    client.detachThingPrincipal(DetachThingPrincipalRequest.builder()
-                            .thingName(thingName())
-                            .principal(p)
-                            .build());
-                });
+        if ((thingName()) != "") {
+            client.listThingPrincipals(ListThingPrincipalsRequest.builder().thingName(thingName()).build()).principals()
+                    .forEach(p -> {
+                        client.detachThingPrincipal(DetachThingPrincipalRequest.builder()
+                                .thingName(thingName())
+                                .principal(p)
+                                .build());
+                    });
 
-        client.deleteThing(DeleteThingRequest.builder()
-                .thingName(thingName())
-                .build());
+            client.deleteThing(DeleteThingRequest.builder()
+                    .thingName(thingName())
+                    .build());
+        }
     }
 }
